@@ -3,6 +3,6 @@ package com.tennecotecnic.onlineshop.model;
 public enum Category {
     BOOK,
     VEHICLE,
-    FOOD;
+    FOOD
 
 }
